pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.129.0",
    "passlib>=1.7.4",
    "psycopg-binary>=3.3.2",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.33",
    "uvicorn>=0.40.0",
]
